requirements.dev.txt
ipython
ipdb
pudb
pytest-watch

# Code quality
flake8
pyproject-flake8

# Install project as editable
-e .
